+++ yash-env/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `Env` struct now contains the `any` field of type `DataSet`.
     - The `DataSet` struct is defined in the newly added `any` module.
       It can be used to store arbitrary data.
+- The `Env` struct now has the `wait_for_subshell_to_halt` method.
+    - This method waits for the subshell to terminate or stop.
 - The `System` trait now has the `get_sigaction` method.
     - This method returns the current signal handling configuration for a signal.
       This method does not modify anything, so it can be used with an immutable